feat(kms): update listing page
ref: MANAGER-13791
